Fix/workloads4intewell

+++ workloads/EPNP/epnp_bench.cpp
@@ -111,8 +111,8 @@ extern "C" int epnp_bench_run(size_t iterations) {
     double avg_time_us = total_time_us / loops;
 
     /* Unified format timing output */
-    printf("[EPNP] samples=%zu total_time=%.3f ms avg_latency=%.3f us/iter\n",
-           loops, total_time_us / 1000.0, avg_time_us);
+    printf("[EPNP] samples=%lu total_time=%.3f ms avg_latency=%.3f us/iter\n",
+           (unsigned long)loops, total_time_us / 1000.0, avg_time_us);
 
     return 0;
 }
